Add implementation for challenge 3 : Largest Prime factor
For this problem, I had to dig a little in my old mathematic's school book.
Well it turns out that all number can be decomposed in a product of prime numbers.
Like : 4872 = 2^3 x 3 x 7 x 29

So I have adopted this approach to solve the problem.
I created a ErathostheneTable class which allows me to have at my disposal
a Sieve of Erathosthene whose size I can configure using a Template class.
After that I apply the algorithm by going down the list of prime numbers
to decompose the indicated value : 600851475143.

- Compilation time : 84ms
- Execution time : 84ms
- Platform : https://godbolt.org/