Add nanobench to project
Nanobench seems to be a very good header files benchmarking library.
At the moment the file is very small 120 ko so I include it in every
project.

But I will start refactoring the whole repository to have only one reference
to nanobench and all project using it (give me some time :) ).

But so far nanobench is good, I will be watching this project very close now.