Add solution for challenge 4
For this solution, I first used a method to determine
when a number is a palindrom using std::to_string to
convert a number into a string and then analyzing
the necessary characters.

For the rest, it was the double loop to find the largest
palindrome products of two 3-digit numbers.